chore: define a JS file for each card, put them into a folder
